+++ CHANGELOG.md
@@ -16,6 +16,9 @@ of response and not the raw API endpoint response
 
 ### Added
 
+- **CUMULUS-2577**
+  - Adds `POST /executions` endpoint to create an execution
+
 - **CUMULUS-2592**
   - Adds logging when messages fail to be added to queue